--- a/logstash-filter-date.gemspec
+++ b/logstash-filter-date.gemspec
@@ -8,10 +8,10 @@ Gem::Specification.new do |s|
   s.authors         = ["Elastic"]
   s.email           = 'info@elastic.co'
   s.homepage        = "http://www.elastic.co/guide/en/logstash/current/index.html"
-  s.require_paths = ["lib"]
+  s.require_paths = ["lib", "vendor/jar-dependencies"]
 
   # Files
-  s.files = Dir['lib/**/*','spec/**/*','vendor/**/*','*.gemspec','*.md','CONTRIBUTORS','Gemfile','LICENSE','NOTICE.TXT']
+  s.files = Dir['lib/**/*','spec/**/*','vendor/**/*','*.gemspec','*.md','CONTRIBUTORS','Gemfile','LICENSE','NOTICE.TXT', "vendor/jar-dependencies/**/*.jar"]
 
   # Tests
   s.test_files = s.files.grep(%r{^(test|spec|features)/})
@@ -25,5 +25,6 @@ Gem::Specification.new do |s|
   s.add_development_dependency 'logstash-codec-json'
   s.add_development_dependency 'logstash-output-null'
   s.add_development_dependency 'logstash-devutils'
+  s.add_development_dependency 'benchmark-ips'
 end