--- a/src/main/java/org/logstash/filters/DateFilter.java
+++ b/src/main/java/org/logstash/filters/DateFilter.java
@@ -1,4 +1,5 @@
-/* * Licensed to Elasticsearch under one or more contributor
+/*
+ * Licensed to Elasticsearch under one or more contributor
  * license agreements. See the NOTICE file distributed with
  * this work for additional information regarding copyright
  * ownership. Elasticsearch licenses this file to you under