--- a/src/main/java/org/logstash/filters/ParseExecutionResult.java
+++ b/src/main/java/org/logstash/filters/ParseExecutionResult.java
@@ -1,0 +1,7 @@
+package org.logstash.filters;
+
+public enum ParseExecutionResult {
+  SUCCESS,
+  FAIL,
+  NO_INPUT_FOUND
+}