… .idea/copyright/Apache_2_0___Elastic.xml
… .idea/copyright/profiles_settings.xml
… .travis.yml
… gradle/wrapper/gradle-wrapper.jar
… lib/logstash/filters/date-old.rb
… spec/filters/date_spec.rb
… spec/fixtures/old_date_filter.rb
… src/main/java/org/logstash/filters/FieldSetter.java
… src/main/java/org/logstash/filters/InputHandler.java
… src/main/java/org/logstash/filters/NumericParserExecutor.java
… src/main/java/org/logstash/filters/ParserExecutor.java
… src/main/java/org/logstash/filters/ResultSetter.java
… src/main/java/org/logstash/filters/StringInputHandler.java
… src/main/java/org/logstash/filters/TextParserExecutor.java
… src/main/java/org/logstash/filters/TimestampSetter.java
… src/main/java/org/logstash/filters/parser/CasualISO8601Parser.java
… src/main/java/org/logstash/filters/parser/JodaParser.java
… src/main/java/org/logstash/filters/parser/TAI64NParser.java
… src/main/java/org/logstash/filters/parser/TimestampParser.java
… src/main/java/org/logstash/filters/parser/TimestampParserFactory.java
… src/main/java/org/logstash/filters/parser/UnixEpochParser.java
… src/main/java/org/logstash/filters/parser/UnixMillisEpochParser.java
… src/test/java/org/logstash/filters/DateFilterTest.java
… src/test/java/org/logstash/filters/parser/JodaParserTest.java
… src/test/java/org/logstash/filters/parser/UnixEpochParserTest.java
… test/perf.rb