TODO items:

* [x] Include #73 - UNIX epoch > 2^31 fails parsing 
* [x] Rebase to make this mergable into master.

-----

My main goal was to resolve #41 (date parse failures causing an unexpected performance hit).

The implementation details are that this plugin is largely written in Java now. I have done my best to make the existing (master) specs pass with the new code.

I believe this PR solves #41:

As a side effect, this PR also dramatically improves performance many date filter configurations.

Benchmarks: [perf.rb](https://gist.github.com/jordansissel/0d537219107ea9f09a2cb7137f241721)

The value below like `70k (± 5%) i/s` is rate -- iterations per second -- and percentage deviation. Higher iteration rate is better.


| input | patterns | old code | new code | ratio |
|-------|----------|----------|----------|-------|
| 2010-01-01T01:01:01Z | ISO8601 |     71.778k (± 5.2%) i/ |    163.712k (± 5.4%) i/ | 2.28 |
| 2010-01-01T01:01:01Z | YYYY, ISO8601 |      3.730k (± 3.9%) i/ |     53.403k (± 3.7%) i/ | 14.32 |
| 2010-01-01T01:01:01Z | YYYY, YYYY-DD, ISO8601 |      2.020k (± 3.3%) i/ |     33.268k (± 3.7%) i/ | 16.47 |
| 1479171046 | UNIX |     87.744k (± 4.2%) i/ |    297.769k (± 5.9%) i/ | 3.39 |
| 1479171070.26 | UNIX |     86.535k (± 3.9%) i/ |    302.593k (± 5.1%) i/ | 3.50 |
| 1479171094338 | UNIX_MS |     88.876k (± 4.0%) i/ |    299.002k (± 5.9%) i/ | 3.36 |
| January  1 2015 | MMMM  d YYYY, MMM dd YYYY |     72.107k (± 4.3%) i/ |    189.660k (± 6.0%) i/ | 2.63 |
| January 11 2015 | MMMM  d YYYY, MMM dd YYYY |      3.762k (± 3.8%) i/ |     53.946k (± 5.4%) i/ | 14.34 |
| Jan  1 2015 | MMM  d YYYY, MMM dd YYYY |     71.691k (± 5.1%) i/ |    189.021k (± 5.7%) i/ | 2.64 |
| Jan 11 2015 | MMM  d YYYY, MMM dd YYYY |      3.856k (± 3.4%) i/ |     53.770k (± 4.2%) i/ | 13.94 |
| [Jan  1 2015, ... 50 more ] | MMM  d YYYY, MMM dd YYYY |      1.464k (± 6.9%) i/ |      4.224k (± 7.0%) i/ | 2.89 |

Test results:

```
% ./gradlew --no-daemon build rubyTest                                                                                                                         
...
Finished in 1.34 seconds (files took 2.48 seconds to load)                                                                                                     
63 examples, 0 failures, 1 pending                                                                                                                             
                                                                                                                                           
Randomized with seed 1
```