@jordansissel 2016-11-07
Awww yeeeah. Travis is green. https://travis-ci.org/logstash-plugins/logstash-filter-date/builds/174004294