@jordansissel 2016-11-09
Merged #73 into this by doing the following:

```
curl -s https://patch-diff.githubusercontent.com/raw/logstash-plugins/logstash-filter-date/pull/73.patch | sed -e 's,/date.rb,/date-old.rb,g' | git am
```

I wonder if this is going to majorly break the branch.. hmm