@suyograo 2016-11-14
@jordansissel there are bunch of commented code and `System.out` code which can be either removed or converted to `logger.debug`.