@jordansissel 2016-11-15
I've included perf.rb now as test/perf.rb.