@jordansissel 2016-11-15
Rebased onto master.