@jordansissel 2016-11-15
@suyograo I think I'm done making changes based on your last round of feedback. Tests are passing for me and I've edited the description to include the latest performance results