@suyograo 2016-11-15
@jordansissel <3 the perf bump from this PR. LGTM.